requirements.txt
mkdocs-gen-files==0.5.0
mkdocs-literate-nav==0.6.2
mkdocs-material==9.6.11
mkdocs-section-index==0.3.10
ipython==9.6.0
mypy==1.18.2
ruff==0.14.0
types-PyYAML==6.0.12.3
PyYAML>=6.0
pytest-benchmark==5.1.0
pytest-cov==7.0.0
pytest==8.4.2
